requirements.txt
click~=7.1.2
pydriller~=1.15.5
terminaltables~=3.1.0
gitpython~=3.1.14
questionary~=1.9.0